+++ finalProject/src/main/java/com/univ/fin/member/controller/AdminController.java
@@ -1,0 +1,53 @@
+package com.univ.fin.member.controller;
+
+import java.sql.Date;
+
+import javax.servlet.http.HttpSession;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import com.univ.fin.member.model.service.MemberService;
+import com.univ.fin.member.model.vo.Student;
+
+@Controller
+public class AdminController {
+	
+	@Autowired
+	private MemberService memberService;
+	
+	@Autowired
+	private BCryptPasswordEncoder bcryptPasswordEncoder;
+	
+	//학생등록 페이지
+	@RequestMapping("enrollStudent.ad")
+	public String enrollStudent() {		
+		
+		return "member/admin/enrollStudent";
+	}
+	
+	//학생등록 페이지 등록
+	@RequestMapping(value="insertStudent.ad" ,method=RequestMethod.POST)
+	public String insertStudent(Student st,
+								Model model,
+								HttpSession session) {
+	System.out.println(st.toString());
+		
+	int result = memberService.insertStudent(st);
+	
+	if(result>0) {
+		model.addAttribute("msg", "회원 생성 완료");
+	}else {
+		model.addAttribute("msg","회원 생성 실패");
+	}
+	return "member/admin/enrollStudent";
+		
+	}
+	
+
+
+}
