+++ finalProject/src/main/java/com/univ/fin/member/controller/ProfessorController.java
@@ -1,4 +1,4 @@
-package com.univ.fin.member.controller;
+ package com.univ.fin.member.controller;
