+++ finalProject/src/main/java/com/univ/fin/member/model/service/MemberService.java
@@ -1,5 +1,6 @@
 package com.univ.fin.member.model.service;
 
+import java.sql.Date;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -83,9 +84,11 @@ public interface MemberService {
 	
 	//학적정보 수정 (교수)
 	int updateProfessor(Professor pr);
-
-	// 회원추가 (학생)
+	
+	//학생추가 (관리자)
 	int insertStudent(Student st);
+
+
